Move TestGenerator into pymeasure folder.
