+++ pymeasure/generator.py
